+++ src/main/java/lotto/domain/LottoNumber.java
@@ -6,7 +6,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import lotto.domain.creator.NumbersCreator;
 
 public final class LottoNumber {
     private static final int MIN_NUMBER = 1;
