+++ src/main/java/lotto/domain/NumbersCreator.java
@@ -1,4 +1,4 @@
-package lotto.domain.creator;
+package lotto.domain;
 
 import java.util.List;
