+++ src/main/java/lotto/domain/winning/BonusMatchStatus.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.winning;
 
 public enum BonusMatchStatus {
     MATCHED {
