+++ src/main/java/lotto/domain/winning/WinningResult.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.winning;
+
+import lotto.domain.Money;
 
 import java.util.Arrays;
