+++ src/test/java/lotto/domain/WinningStatisticsTest.java
@@ -5,6 +5,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
+
+import lotto.domain.winning.WinningNumbers;
+import lotto.domain.winning.WinningResult;
+import lotto.domain.winning.WinningStatistics;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
